@@ -5,7 +5,7 @@ repository: https://github.com/aikalant/dart-custom-analyzer-plugin
 publish_to: none
 
 environment:
-  sdk: '>=2.14.0-0 <3.0.0'
+  sdk: '>=2.17.0-0 <3.0.0'
 
 dependencies:
    path: ^1.8.0
@@ -17,3 +17,6 @@ dependencies:
 
 dev_dependencies:
   lints: ^1.0.0
+  mocktail: ^0.3.0
+  recase: ^4.1.0
+  test: ^1.20.2
