@@ -4,6 +4,7 @@ import 'package:analyzer/src/analysis_options/analysis_options_provider.dart';
 import 'package:analyzer/src/dart/analysis/driver.dart';
 import 'package:glob/glob.dart';
 import 'package:path/path.dart' as p;
+
 import 'yaml_utils.dart';
 
 Map<String, RuleConfig> getOptionsFromDriver(
@@ -39,13 +40,14 @@ Map<String, RuleConfig> _parseOptions(
   final ruleConfigs = <String, RuleConfig>{};
   final rootNode = options['custom_linter'];
   if (rootNode is Map<String, Object>) {
-    final defaultUrl =
-        rootNode['url'] is String ? rootNode['url'] as String? : null;
+    final defaultUrl = rootNode['documentationUrl'] is String
+        ? rootNode['documentationUrl'] as String?
+        : null;
     final rulesNode = rootNode['rules'];
     if (rulesNode is List<Object>) {
       for (final ruleNode in rulesNode) {
         if (ruleNode is String) {
-          ruleConfigs[ruleNode] = RuleConfig(defaultUrl);
+          ruleConfigs[ruleNode] = RuleConfig(documentationUrl: defaultUrl);
         } else if (ruleNode is Map<String, Object>) {
           final entry = ruleNode.entries.first;
           final ruleID = entry.key;
@@ -53,7 +55,10 @@ Map<String, RuleConfig> _parseOptions(
           if (ruleOptions is Map<String, Object>) {
             _applyOptionsToConfig(
               ruleOptions,
-              ruleConfigs.putIfAbsent(ruleID, () => RuleConfig(defaultUrl)),
+              ruleConfigs.putIfAbsent(
+                ruleID,
+                () => RuleConfig(documentationUrl: defaultUrl),
+              ),
               root,
             );
           }
@@ -74,9 +79,9 @@ void _applyOptionsToConfig(
     config.enabled = enabled;
   }
 
-  final url = options['url'];
+  final url = options['documentationUrl'];
   if (url is String) {
-    config.url = url;
+    config.documentationUrl = url;
   }
 
   final excludedGlobs = options['exclude'];
@@ -94,10 +99,25 @@ void _applyOptionsToConfig(
   );
 }
 
+/// Configuration provided in a client project `analysis_options.yaml` under the
+/// `custom_linter` key.
 class RuleConfig {
-  RuleConfig(this.url);
+  /// Creates a new rule configuration with the given properties.
+  RuleConfig({this.documentationUrl});
+
+  /// Whether or not the rules in this package are enabled.
   bool enabled = true;
-  String? url;
+
+  /// {@template documentation_url}
+  /// Optional path to a markdown file describing the rule. The markdown file
+  /// must contain a heading with the same name as the rule id.
+  /// {@endtemplate}
+  String? documentationUrl;
+
+  /// Glob exclusion patterns which prevent analysis from occurring by the
+  /// rules in this package.
   List<Glob> excludedGlobs = [];
+
+  /// Individual rule options.
   Map<String, Object> options = {};
 }
