@@ -2,27 +2,30 @@ import 'package:analyzer/dart/analysis/results.dart';
 import 'package:analyzer_plugin/protocol/protocol_common.dart';
 
 import '../../options/options.dart';
-import '../rule_base.dart';
-import '../visitor_mixin.dart';
-import '_visitor.dart';
+import '../../rule/rule.dart';
+import 'no_material_cupertino_imports_visitor.dart';
 
 class NoMaterialCupertinoImportsRule extends Rule {
   @override
   String get id => 'no_material_cupertino_imports';
 
   @override
-  String get message => 'Do not use material or cupertino libraries.';
+  String get message => "Don't use material or cupertino libraries.";
 
   @override
-  String? get correction => "Instead, use 'package:flutter/widgets.dart'";
+  String? get correction => "Import 'package:flutter/widgets.dart' instead.";
 
   @override
   AnalysisErrorSeverity get severity => AnalysisErrorSeverity.ERROR;
 
   @override
-  VisitorMixin getVisitor(
+  SimpleRuleVisitor getVisitor(
     ResolvedUnitResult result,
-    RuleConfig options,
+    RuleConfig config,
   ) =>
-      Visitor(this, result, options);
+      NoMaterialCupertinoImportsVisitor(
+        rule: this,
+        result: result,
+        config: config,
+      );
 }
